pyproject.toml
[project]
name = "natural-language-fabric"
version = "0.1.0"
description = "Using LLMs to solve genetic optimization problems"
authors = [{ name = "Gerhard Stenzel", email = "gerhard.stenzel@ifi.lmu.de" },]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "jinja2>=3.1.5",
    "openai>=1.59.8",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
    "jupyter>=1.1.1",
    "ruff>=0.9.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "pandas>=2.2.3",
    "orjson>=3.10.14",
    "ollama>=0.4.7",
    "scikit-learn>=1.6.1",
    "fastparquet>=2024.11.0",
]


[tool.ruff]
line-length = 110

[tool.pyright]
exclude = ["build", "dist", ".venv", ".ipynb_checkpoints", ".pytest_cache", ".ruff_cache", ".git"]
